refactor: 비교 완료 페이지를 (stack) 라우트 그룹으로 이동 및 views 레이어 분리
