+++ apps/knockdog/app/(stack)/compare-complete/page.tsx
@@ -1,0 +1,10 @@
+import { CompareCompletePage } from '@views/compare-complete-page';
+import { Suspense } from 'react';
+
+export default function Page() {
+  return (
+    <Suspense>
+      <CompareCompletePage />
+    </Suspense>
+  );
+}
