+++ apps/knockdog/src/views/compare-complete-page/index.ts
@@ -1,0 +1,1 @@
+export { CompareCompletePage } from './ui/CompareCompletePage';
